Files created while learning the RRT algorithm in March
This uses pygame to implement the RRT algorithm. The path computed sometimes goes through the obstacles. Hence we need to use shapely to eliminate such paths. Also when a narrow path is present between the start and goal, the tree of nodes grow too big and crashes. This could be eliminated by getting proper bias taking in account the geometry of the obstacles and appropriately finding the optimal node.